added the login and signup
